Fixed a bug with the PropertyType's config object
We needed to make a copy of the config object, as previously the object reference was being passed in.
Meaning that any changes made to the config in the editor's controller would affect the subsequent editors loaded in.

NOTE: I'm using JSON.parse/stringify, as that's the quickest way I know of to deep-copy a JS object.